# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OrbChatInput {
  /** @minLength 1 */
  message: string;
  /** @nullable */
  conversationId?: string | null;
  /** @nullable */
  chakra?: string | null;
}

export interface OrbChatReply {
  reply: string;
  /** @nullable */
  conversationId?: string | null;
}

export interface InquiryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 3 */
  email: string;
  /** @minLength 1 */
  message: string;
  source?: string;
  /** @nullable */
  conversationId?: string | null;
}

export interface Inquiry {
  id: number;
  name: string;
  email: string;
  message: string;
  source: string;
  /** @nullable */
  conversationId?: string | null;
  createdAt: string;
}