# lib/api-zod/src/generated/types/inquiry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Inquiry {
  id: number;
  name: string;
  email: string;
  message: string;
  source: string;
  /** @nullable */
  conversationId?: string | null;
  createdAt: Date;
}