# lib/api-zod/src/generated/types/inquiryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InquiryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 3 */
  email: string;
  /** @minLength 1 */
  message: string;
  source?: string;
  /** @nullable */
  conversationId?: string | null;
}