# lib/api-zod/src/generated/types/orbChatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrbChatInput {
  /** @minLength 1 */
  message: string;
  /** @nullable */
  conversationId?: string | null;
  /** @nullable */
  chakra?: string | null;
}