# lib/api-zod/src/generated/types/orbChatReply.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrbChatReply {
  reply: string;
  /** @nullable */
  conversationId?: string | null;
}